+++ application/models/Images.php
@@ -5,7 +5,8 @@ class Images extends CI_Model {
     /**
      * Constructor for Images class.
      */
-    function __contruct() {
+    function __contruct() 
+    {
         parent::__construct();
     }
 
@@ -13,7 +14,8 @@ class Images extends CI_Model {
      * Queries the database for images and returns the result
      * as an array. The images are sorted by ID in descending order.
      */
-    function all() {
+    function all() 
+    {
         $this->db->order_by("id", "desc");
         $query = $this->db->get('images');
         return $query->result_array();
@@ -22,7 +24,8 @@ class Images extends CI_Model {
     /**
      * Queries the database for the newest images.
      */
-    function newest() {
+    function newest()
+    {
         $this->db->order_by("id", "desc");
         $this->db->limit(3);
         $query = $this->db->get('images');