+++ changelog.md
@@ -14,6 +14,8 @@ Release Date: Jan 25, 2017
 
     -   Gallery (Will Murphy)
 
+    -   Welcome (Will Murphy)
+
 -   views
 
     -   gallery (Will Murphy)